Created Unit Tests for Limit and toDifferentialDrive Methods
The test for limits makes sure the rotational drive signals are never above 1 or below -1.
All limit test methods produced no errors.
The toDifferentialDriveSignal tests to make sure that motor outputs are not above 1 or below -1.
The limit method is never implemented to limit motor outputs. Test results come up with motor outputs of up to 2 and -2.