Limit Values
This addition limits the values we give the motors on either sides to address the flawed input system we currently have. We will still turn, and the rotation will still work, but we now actually give the motor controllers legal values between 1 and -1. The old system still worked because the motor controllers did this for us automatically.