Make Changes
I made my own version of the control scheme that I can play with and work on. I added getters to get the two buttons on top that I want to know about, and I also added constants for the deadband functions. I added the deadband functions in rotational drive signal and therefore referencing the DriveSignalUtilities class. I fixed an error where a reference was made to the limit in the actual robot.java class, and moved it to where it was supposed to be in DrivePipeline. I added the deadband function to the pipeline, where it can get called.