Make My ControlScheme
I created a new controlScheme that utilized the easily accessible buttons on the top of the joystick controller to swivel the robot at the drive limit speed. That limit can be changed in the future. I also changed the robot.java to use my new control scheme. This new scheme still has all uses and methods as the defaultControlScheme we use, just with additional code to detect buttons and such.