Latest Version of Files
The Path.java and PathDriver.java files have been updated to what is currently developed. AutoCSVReader.java has also been added. AutoCSVReader.java is run on the robot's side of things in order to read setpoints out of a .csv file on the roboRIO. These setpoints are then given to the driving code on the robot for autonomous.